fixed import statements after moving components
